Add refs.Container and refs.Update
Consider two changes: set a ref path to be [x, 5] and insert an item
at index zero into arry at [x]

If these were done with regular change.Replace and change.Splice,
it would appear as if the two changes do not conflict.  It would
cause the ref path to be [x, 5] instead of [x, 6].

The Update operation is a special version of changes.Replace for
setting references and it correctly handles the merge.

The existing composition techniques with change.PathChange forces
a constraint on Update: It can only be merged with other changes
at the same "path".  So, if an Update refers to a path outside
(such as ../z/5) and changes happen on that path, it wont be properly
merged with it.

refs.Container works around this constraint by requiring that all
references are effectively to its own children. It wraps an arbitrary
value and also exposes ability to manage references within.

The main initial use case is streams/text/Editable which would need
to a set of "selections" to regions of the text.