ux: fn: Add a functional UX building experiment
Unlike the  <simple> mechanism, components are expected  to be functions:

```golang
func MyComponent(c *myContext, arg1 Arg1) core.Element {
   return fn.Element(
       key,
       c.SubComponent(key, args...),
       ...
   )
}
```

In this example, the component is *MyComponent* and it takes a context.
The context type is auto-generated.

The component calls a sub-component via the context: *c.SubComponent(..)*

The code generator ensures that the context  has a method SubComponent
which maps to thee SubComponent(...) call (substituting the key for the
corresponding context maintained internally).

The PR includes a todo.go which illustrates both how the actual code can
be written and how the generated code looks like.  This example also shows
how all this syntactic sugar preserves the bridge to regular <simple>
components.