Migrate all reactive UX related stuff to github.com/dotchain/fuss
The version up at fuss is a much cleaned up and more coherent implementation.

The main motivation is the fact that dot/ has a substantially different flavor than the  strongly-typed  code-generation based reactive implementation.

The two projects can probably converge but it is better to keep these separate while the issues are sorted out.