cleanup interop tests to match dataset repo
The changes to merge are minimal and make it actually a little better in some odd cases.

This allows both the JS and Go version to be in sync better by comparinng against the master json tests.

The dotchain/dataset repo has been updated as well.

For all practical purposes, the merge behavior will likely be frozen onces the JS implementation conforms to the same set which is likely very soon.