Add fred: a functional reactive engine
Fred is an experimental functional reactive engine on top of DOT.  Note that this is a "weakly typed" setup (where the dynamically created expressions are typeless and interpreted).  A strongly typed setup will very likely approach from the language side rather than ground up but this is easier and useful in its own right. The plan is to learn from this and then figure out how to properly divide boundaries and get type safety.

1. Dir holds directory of object defintions
2. Values can be evaluated.  All evaluations are *live* using a stream-like interface
3. Recalcuations are pull-based, so items not *pulled* will generally incur no cost
4. Cache of common values and expressions to avoid recomputation.

The directory only implements error values and references right now. Rich types are
not difficult but the motivation for this experiment is mostly to see how to sensibly
editing calculated values.  For instance, what does it mean to edit the value of a reference?

TODO

[ ] Add ability to get provenance of some path in a value (i.e. where this comes from originally).  The provenance for data is like stack trace for code on steroids.
[ ] Add ability to mutate computed values.  The implementation is similar to provenancee but likely not the same for computed values.
[ ] Add simple types (number, sum). These will likely be needed for illustrating the interesting parts of the above.
[ ] Add generic functions (worth converting sum to a generic function)
[ ] Add container types
[ ] Add type conversion heuristics