Add fred (second attempt)
Fred is an experimental functional reactive engine on top of DOT.  Note that this is a "weakly typed" setup (where the dynamically created expressions are typeless and interpreted).  A strongly typed setup will very likely approach from the language side rather than ground up but this is easier and useful in its own right. The plan is to learn from this and then figure out how to properly divide boundaries and get type safety.

1. A Def holds something that can be computed (these have an Eval method)
2. A Val expresses a fixed computed value
3. Env captures the environment needed for a single Def computation
    * A resolver which resolves references (needed for computation)
    * A cache holds cached results to make shared sub-expressions evaluate faster
    * A recursion checker to help detect recursive evaluations

There are other helpers defined (such as Scope for managing scopes etc)

TODO:

[ ] Add simple types and type conversion heuristics
[ ] Add Refs, Filter, Map, Reduce
[ ] Add streaming ability to get the next value of a computation without recomputing everything
[ ] Add provenance of parts of derived values (check where a particular value comes from)
[ ] Add edit of parts of derived values