add rich text data type
This is derived from several earlier versions that moved around.  Not sure about x/ location but will probably keep it there until I'm happy with the full interface.

Lists and other block formatting are intended to be stored as sub-nodes (via the "Embed" attribute). This does severely compromise merge behavior (converting a region to text will not merge well with someone editing that region) but makes the code and structure a lot simpler and natural.

All the specific attributes are currently in rich/html though it is unclear if those belong there as it would be nice to have the same attributes render into different formats (such as markdown).