WIP: Initial draft of rich text editor model
This just sets up the structure of Editor and its associated Stream.

The plan is to add logical navigation controls and sensible edit behaviors (such as tab indenting within lists etc) here.

This will not include visual elements (such as visual "up" navigation or scroll preservation).

Collaborative cursors do not have a clear home at this point but will likely atleast include a special Region or Selection attribute to deal with styling.

It is unclear if this should be used for browser rendering or if it should be ported to ES6 but deferring that until later -- porting from Go to ES6 is rather straightforward and can possibly even be automated for code like this.